Add context component to handle context providers and state
